Move psd_tools.api.composer module to psd_tools.composer subpackage
