+++ src/psd_tools/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import, unicode_literals
 from .api.psd_image import PSDImage
-from .api.composer import compose
+from .composer import compose
